% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.BIN.R
\docType{data}
\name{Dados.BIN}
\alias{Dados.BIN}
\title{Dados de exemplo (Dados binarios).}
\format{
Um data frame com 8 observacoes e 9 variaveis.
}
\usage{
data(Dados.BIN)
}
\description{
Exemplo simulado com um conjunto de dados binarios.
}
\references{
PlayList "Curso de Analise Multivariada":
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
