% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.CAT.R
\docType{data}
\name{Dados.CAT}
\alias{Dados.CAT}
\title{Exemplo de dados qualitativos multicategoricos.}
\format{
Um data frame com 9 observacoes e 13 variaveis.
}
\usage{
data(Dados.CAT)
}
\description{
Exemplo  com um conjunto de dados com 13 variaveis resposta do
tipo multicategorico para 9 individuos.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
