% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.DBC.Misto.R
\docType{data}
\name{Dados.DBC.Misto}
\alias{Dados.DBC.Misto}
\title{Dados de exemplo de um experimento em DBC com dados quantitativos e mistos.}
\format{
Um data frame. A primeira coluna
a identificacao dos tratamentos, a segunda  a
identificacao de blocos e da terceira em diante as variaveis respostas
 (dependentes).
}
\usage{
data(Dados.DBC.Misto)
}
\description{
Exemplo de parte de um exeperimento com clones batata-doce em DBC
com dados quantitativos e qualitativos.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
