% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.DBC.R
\docType{data}
\name{Dados.DBC}
\alias{Dados.DBC}
\title{Dados de exemplo de um experimento em DBC.}
\format{
Um data frame com 45 observacoes e 11 variaveis. A primeira coluna
a identificacao dos tratamentos, a segunda  a
identificacao de blocos e da terceira em diante as variaveis respostas
 (dependentes).
}
\usage{
data(Dados.DBC)
}
\description{
Exemplo simulado com um conjunto de dados
de um experimento em DBC.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
