% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.DIC.R
\docType{data}
\name{Dados.DIC}
\alias{Dados.DIC}
\title{Dados de exemplo de um experimento em DIC.}
\format{
Um data frame com 30 observacoes e 9 variaveis. A primeira coluna
tem a identificacao dos tratamentos, a segunda  a
identificacao das repeticoes e da terceira em diante as variaveis respostas
 (dependentes).
}
\usage{
data(Dados.DIC)
}
\description{
Exemplo simulado com um conjunto de dados
de um experimento no delineamento inteiramente casualizado
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
