% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.DQL.R
\docType{data}
\name{Dados.DQL}
\alias{Dados.DQL}
\title{Dados de exemplo de um experimento em DQL.}
\format{
Um data frame com 16 observacoes e 7 colunas. A primeira coluna
tem a identificacao dos tratamentos, a segunda um vetor com a
identificacao das linhas, a terceira coluna com a identificacao das colunas
e da quarta em diante as variaveis respostas (dependentes).
}
\usage{
data(Dados.DQL)
}
\description{
Exemplo simulado com um conjunto de dados
de um experimento em quadrados latino
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
