% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.Fat2.DIC.R
\docType{data}
\name{Dados.Fat2.DIC}
\alias{Dados.Fat2.DIC}
\title{Dados de um experimento em esquema fatorial duplo conduzido em DIC.}
\format{
Um data frame com 36 observacoes e 7 colunas. A primeira coluna
com a identificacao do fator A, a segunda  com a
identificacao do fator B, a terceira coluna com a identificacao das repeticoes
e da quarta em diante as variaveis respostas (dependentes).
}
\usage{
data(Dados.Fat2.DIC)
}
\description{
Exemplo  com um conjunto de dados
de um experimento em esquema fatorial duplo conduzido em DIC.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
