% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.Fat3.DBC.R
\docType{data}
\name{Dados.Fat3.DBC}
\alias{Dados.Fat3.DBC}
\title{Dados de um experimento em esquema fatorial triplo conduzido em DBC.}
\format{
Um data frame com 64 observacoes e 9 colunas. A primeira coluna
tem a identificacao do fator A, a segunda  com a
identificacao do fator B, a terceira coluna com a identificacao do fator C,
 a quarta coluna com a identificacao dos blocos
e da quinta em diante as variaveis respostas (dependentes).
}
\usage{
data(Dados.Fat3.DBC)
}
\description{
Exemplo  com um conjunto de dados
de um experimento em esquema fatorial duplo conduzido no delineamento em
blocos casualizados.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
