% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.FMI.Quali.R
\docType{data}
\name{Dados.FMI.Quali}
\alias{Dados.FMI.Quali}
\title{Exemplo de dados qualitativos multicategoricos em um experimento com progenies
de meios irmaos.}
\format{
Um data frame com 25925 observacoes e 10 variaveis.
}
\usage{
data(Dados.FMI.Quali)
}
\description{
Exemplo  de um experimento com progenies de meios irmaos de couve
}
\references{
Brito et al. (2021). Genetic divergence between half-sibling progenies of
 kale using different multivariate approaches. Horticultura Brasileira 39:
  178 185. <http://dx.doi.org/10.1590/s0102 0536 20210208>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
