% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.MED.R
\docType{data}
\name{Dados.MED}
\alias{Dados.MED}
\title{Dados de um exemplo com variaveis quantitativas.}
\format{
Um data frame com 10 observacoes e 7 variaveis.
}
\usage{
data(Dados.MED)
}
\description{
Exemplo  com um conjunto de dados com 7 variaveis resposta
quantitativas para 10 individuos.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
