% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.Misto.R
\docType{data}
\name{Dados.Misto}
\alias{Dados.Misto}
\title{Dados de exemplo de um experimento  com dados mistos.}
\format{
Um data frame com 5 observacoes e 11 colunas.
}
\usage{
data(Dados.Misto)
}
\description{
Exemplo simulado com um conjunto de dados
de um experimento com dados qualitativos binarios, multicategoricos
e quantitativos.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
