% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dendograma.R
\name{Dendograma}
\alias{Dendograma}
\title{Funcao substituida para o nome correto (Dendrograma)}
\usage{
Dendograma(...)
}
\arguments{
\item{...}{Qualquer parametro.}
}
\description{
Esta funcao teve seu nome alterado para Dendrograma
}
\seealso{
\code{\link{Dendrograma}}
}
