% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateAnalysis-package.R
\name{MultivariateAnalysis package}
\alias{MultivariateAnalysis package}
\alias{MultivariateAnalysis}
\title{Pacote Para Analise De Experimentos Com Abordagem Multivariada}
\description{
Este pacote realiza a analise de experimentos
 por abordagem multivariada. Com este pacote e possivel
 fazer analises de variancia multivariada, estimar medidas de
 disimilaridade, fazer dendogramas,componentes principais,
 variaveis canonicas, etc.
}
\references{
PlayList "Curso de Analise Multivariada":
 https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR


CRUZ, C.D. and CARNEIRO, P.C.S.  Modelos biometricos aplicados ao
  melhoramento genetico. 3nd Edition. Vicosa, UFV, v.2, 2014. 668p.  (ISBN: 8572691510)

FERREIRA, D.F. Estatistica Multivariada. (2018) 3ed. UFLA. 624p. (ISBN 13:978 8581270630)

 HAIR, J.F. Multivariate Data Analysis.  (2016) 6ed. Pearson Prentice HalL.
  (ISBN 13:978 0138132637)
}
\author{
Alcinei Mistico Azevedo: <alcineimistico@hotmail.com>
}
