% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariaveisCanonicas.R
\name{VariaveisCanonicas}
\alias{VariaveisCanonicas}
\title{Analise de variaveis canonicas}
\usage{
VariaveisCanonicas(Dados,Modelo,xlab="VC 1", ylab="VC 2",CR=TRUE,
 CorPlot=TRUE, bty="L",Perc)
}
\arguments{
\item{Dados}{Matriz contendo os dados para execucao da MANOVA. Para cada
modelo o conjunto de dados precisa estar organizado de uma forma
apropriada:
 \itemize{
\item Modelos 1 e 2: As duas primeiras colunas devem conter a
identificacao dos tratamentos, e as demais os valores observanos nas
variaveis respostas.
\item Modelo 3: As tres primeiras colunas devem conter as
informacoes dos tratamentos, linhas e colunas, e posteriormente, os valores
da variavel resposta.
\item Modelos 4 e 5: as primeiras colunas precisam ter a
informacao do fator A, fator B, repeticao/bloco, e posteriormente, as
variaveis respostas.
}}

\item{Modelo}{Valor numerico indicando o delineamento:
 \itemize{
\item 1 = Delineamento inteiramente casualizado (DIC)
\item 2 = Delineamento em blocos casualizados (DBC)\item 3 = Delineamento em quadrado
  latino (DQL)
  \item 4 = Esquema fatorial em DIC
  \item 5 = Esquema fatorial em DBC
  }}

\item{xlab}{nome do eixo x do grafico de variaveis canonicas}

\item{ylab}{nome do eixo y do grafico de variaveis canonicas}

\item{CR}{Valor logico (TRUE ou FALSE) indicando se aparecera no grafico
a contriuicao relativa de cada eixo.}

\item{CorPlot}{Valor logico. Se for TRUE sera apresentado no grafico as
correlacoes.}

\item{bty}{deve receber un character indicando o tipo de borda desejado no
grafico.
 \itemize{
   \item"o": Todas as bordas
   \item"n": Sem bordas
   \item"7": Acima e a direita
   \item"L": Abaixo + esquerda (Default)
   \item"C": Acima + Direita + Abaixo
   \item"U": Direita + Abaixo + Direita
   }}

\item{Perc}{Valor entre 0 e 1 indicando o recuo dos eixos.}
}
\value{
A funcao retorna resultados associados as variaveis canonicas.
}
\description{
Esta funcao faz a analise dos dados pelo metodo de variaveis
  canonicas.
}
\examples{
\donttest{
#Delineamento inteiramente casualizado (DIC)
data("Dados.DIC")
MANOVA(Dados.DIC,1)
VariaveisCanonicas(Dados.DIC,1)
#Delineamento em blocos casualizados (DBC)
data(Dados.DBC)
MANOVA(Dados.DBC,2)
VariaveisCanonicas(Dados.DBC,2)
#Delineamento em quadrado latino (DQL)
data(Dados.DQL)
MANOVA(Dados.DQL,3)
VariaveisCanonicas(Dados.DQL,3)
#Esquema fatorial em DIC
data(Dados.Fat2.DIC)
MANOVA(Dados.Fat2.DIC,4)
VariaveisCanonicas(Dados.Fat2.DIC,4)
#Esquema fatorial em DBC
data(Dados.Fat2.DBC)
}
}
\references{
PlayList "Curso de Analise Multivariada":
 https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR


CRUZ, C.D. and CARNEIRO, P.C.S.  Modelos biometricos aplicados ao
  melhoramento genetico. 3nd Edition. Vicosa, UFV, v.2, 2014. 668p.  (ISBN: 8572691510)

FERREIRA, D.F. Estatistica Multivariada. (2018) 3ed. UFLA. 624p. (ISBN 13:978 8581270630)

 HAIR, J.F. Multivariate Data Analysis.  (2016) 6ed. Pearson Prentice HalL.
  (ISBN 13:978 0138132637)
}
\seealso{
/code{/link{lm}}, /code{/link{manova}}
}
