% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados.BIN.R
\docType{data}
\name{Dados.BIN}
\alias{Dados.BIN}
\title{Dados de exemplo (Dados binarios).}
\format{
Um data frame com 5 observacoes e 9 variaveis (Cruz et al. 2014).
}
\usage{
data(Dados.BIN)
}
\description{
Exemplo com um conjunto de dados binarios.
}
\references{
PlayList "Curso de Analise Multivariada":
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>

Cruz et al., 2014. Modelos biometricos aplicados ao melhoramento de genetico.
Vol3. editora UFV. 668p.
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
